Move calculate_selection to InteractionType
It makes more sense there.
